// frontend/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type Body_submit_job_api_jobs_post = {
    file: (Blob | File);
};

export type HTTPValidationError = {
    detail?: Array<ValidationError>;
};

export type JobCreatedResponse = {
    job_id: string;
    status?: JobStatus;
};

export type JobExtra = {
    filename?: (string | null);
    ocr_path?: (string | null);
    result_path?: (string | null);
    artifacts_dir?: (string | null);
    image_paths?: Array<(string)>;
    visualisations?: Array<(string)>;
    problem_visualisations?: Array<(string)>;
    [key: string]: unknown | string;
};

export type JobResultResponse = {
    job_id: string;
    result: unknown;
};

export type JobStatus = 'queued' | 'processing' | 'ocr_running' | 'ocr_complete' | 'pipeline_running' | 'completed' | 'failed' | 'unknown';

export type JobStatusResponse = {
    job_id: string;
    status: JobStatus;
    filename?: (string | null);
    created_at?: (string | null);
    updated_at?: (string | null);
    result_path?: (string | null);
    message?: (string | null);
    extra?: JobExtra;
};

export type ValidationError = {
    loc: Array<(string | number)>;
    msg: string;
    type: string;
};

export type ListJobStatusesApiJobsGetData = {
    limit?: number;
    status?: (JobStatus | null);
};

export type ListJobStatusesApiJobsGetResponse = (Array<JobStatusResponse>);

export type SubmitJobApiJobsPostData = {
    formData: Body_submit_job_api_jobs_post;
};

export type SubmitJobApiJobsPostResponse = (JobCreatedResponse);

export type JobStatusApiJobsJobIdGetData = {
    jobId: string;
};

export type JobStatusApiJobsJobIdGetResponse = (JobStatusResponse);

export type JobResultApiJobsJobIdResultGetData = {
    jobId: string;
};

export type JobResultApiJobsJobIdResultGetResponse = (JobResultResponse);

export type DownloadArtifactsApiJobsJobIdDownloadGetData = {
    jobId: string;
};

export type DownloadArtifactsApiJobsJobIdDownloadGetResponse = (unknown);